Update - toko

+++ app/Http/Controllers/Admin/Toko/TokoController.php
@@ -28,8 +28,8 @@ class TokoController extends Controller
     }
 
     public function store(Request $request){
-        $form_user   = $request->input();
-        $form_toko   = $request->input();
+        $form_user   = $request->all();
+        $form_toko   = $request->all();
         $rules_user  = User::$validation_rules;
         $rules_toko  = Toko::$validation_rules;
 
@@ -43,14 +43,30 @@ class TokoController extends Controller
             return redirect('/ck-admin/toko/tambah');
         }
         unset($form_user['password_konfirmasi'], $form_user['kota'], $form_user['siup'], $form_user['npwp'], $form_user['no_rek'], $form_user['nama_bank']);
-        unset($form_toko['password_konfirmasi'], $form_toko['name'], $form_toko['username'], $form_toko['email'], $form_toko['password'], $form_toko['alamat'], $form_toko['no_telp']);
+        unset($form_toko['password_konfirmasi'], $form_toko['name'], $form_toko['username'], $form_toko['email'], $form_toko['password'], $form_toko['alamat'], $form_toko['no_telp'], $form_toko['gambar']);
 
         $form_user['role_id']   = 2;
         $form_user['status_active'] = 1;
-        $form_user['gambar']    = 'image.jpg';
+
+        if (@$form_user['gambar'] != '') {
+            $extension  = $request->file('gambar')->getClientOriginalExtension();
+
+            if ($extension == 'jpg' || $extension == 'png'  || $extension == 'jpeg'
+                || $extension == 'svg' || $extension == 'gif' ) {
+
+                $name   = $form_user['name'] . date('-d_m_Y-h_i_s');
+                $request->file('gambar')->move('img/users/', $name . '.' . $extension);
+                $form_user['gambar']    = $name. '.' . $extension;
+            } else {
+                $request->session()->flash('message', GeneralFunction::$IMAGE_NOT_VALID_MESSAGE);
+                return redirect('/ck-admin/toko/tambah');
+            }
+        } else {
+            $form_user['gambar']    = '';
+        }
 
         $create_user    = app('users')->create($form_user, $rules_user);
-        $create_toko    = [];
+
         if ($create_user) {
             $form_toko['users_id']  = $create_user['data']['id'];
             $form_toko['nama']      = $form_user['name'];
@@ -85,8 +101,13 @@ class TokoController extends Controller
 
 
     public function delete(Request $request, $id) {
-        $delete = app('toko')->delete($id);
-        $request->session()->flash('message', $delete['message']);
+        $toko           = Toko::where('id', $id)->first();
+        $delete_cascade = $toko->users()->delete();
+
+        if ($delete_cascade) {
+            $delete         = app('toko')->delete($id);
+            $request->session()->flash('message', $delete['message']);
+        }
 
         return redirect('/ck-admin/toko/');
     }
